requirements.txt
Flask
line-bot-sdk
discord.py
requests
openai
python-dotenv